requirements.txt
Flask==3.1.2
flask-cors==6.0.1
google-auth==2.43.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.187.0
requests==2.32.5
deepface==0.0.95
numpy==2.1.3
Pillow==12.0.0
opencv-python==4.12.0.88

244497833082-stkji8qg7cb01itn8nsdme6qao8r4qov.apps.googleusercontent.com